pyproject.toml
[project]
name = "trend-scope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.31.0",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.8",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "langgraph-cli[inmem]>=0.3.6",
    "langsmith>=0.4.8",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "pywinauto>=0.6.9",
    "rich>=14.1.0",
    "tavily-python>=0.7.11",
]
